requirements.txt
openai
pillow
numpy
pillow-heif
replicate
